Prioritise reading text from value then label
In some cases, (e.g.: `UILabel`), the `accessibilityValue` of the element is not set automatically.
For those, the `label` property is being used to determine the text being displayed.

If the `accessibilityLabel` on those are customised, it is advised to also set the `accessibilityValue` to the actual text so that the UI tests are able to verify it.