Test stepper using updated button index
We **CANNOT** access the stepper component by setting an `accessibilityIdentifier` and retrieving it.
This means we have to use the buttons directly as opposed to using the stepper.

We should be aware of other buttons in the view and other steppers when calculating the index we are looking to retrieve.

Please note: the index of the steppers are **backwards**, i.e.: the increment index comes before the decrement one.