Filter buttons before attempting to find the stepper ones
A stepper is not accessible by default. Making them accessible is also not a good solution
as it would mean they would be exposed to the user as an accessible element (which it is not).

The current solution aims at finding the buttons with an `accessibilityLabel` of "Increment" or "Decrement" and
then find the right one using the index of the stepper.